% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_constraints.R
\name{apply_calipers}
\alias{apply_calipers}
\title{Apply caliper constraints}
\usage{
apply_calipers(cost_matrix, left, right, calipers, vars)
}
\description{
Calipers impose per-variable maximum absolute differences.
}
\keyword{internal}
