% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_constraints.R
\name{apply_max_distance}
\alias{apply_max_distance}
\title{Apply maximum distance constraint}
\usage{
apply_max_distance(cost_matrix, max_distance = Inf)
}
\description{
Apply maximum distance constraint
}
\keyword{internal}
