% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_blocks.R
\name{assign_blocks_cluster}
\alias{assign_blocks_cluster}
\title{Assign blocks using clustering}
\usage{
assign_blocks_cluster(left, right, block_vars, method, n_blocks, ...)
}
\description{
Assign blocks using clustering
}
\keyword{internal}
