% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_blocks.R
\name{assign_blocks_group}
\alias{assign_blocks_group}
\title{Assign blocks based on grouping variable(s)}
\usage{
assign_blocks_group(left, right, block_by)
}
\description{
Assign blocks based on grouping variable(s)
}
\keyword{internal}
