% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_constraints.R
\docType{data}
\name{BIG_COST}
\alias{BIG_COST}
\title{Large value for forbidden pairs}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
BIG_COST
}
\description{
Large value for forbidden pairs
}
\keyword{internal}
