% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_distance.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{Compute pairwise distance matrix}
\usage{
compute_distance_matrix(left_mat, right_mat, distance = "euclidean")
}
\description{
Compute pairwise distance matrix
}
\keyword{internal}
