% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_core.R
\name{detect_blocking}
\alias{detect_blocking}
\title{Detect and validate blocking}
\usage{
detect_blocking(left, right, block_id, ignore_blocks)
}
\description{
Detect and validate blocking
}
\keyword{internal}
