% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_tiling.R
\name{.benchmark_square_tiling}
\alias{.benchmark_square_tiling}
\title{Compare performance: hierarchical vs square tiling}
\usage{
.benchmark_square_tiling(
  A_planar,
  B_planar,
  H,
  W,
  max_patch_size = 3L,
  alpha = 1,
  beta = 0.1
)
}
\description{
Compare performance: hierarchical vs square tiling
}
