% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.downscale_both}
\alias{.downscale_both}
\title{Compute downscaled images, return both originals and downscaled with dims}
\usage{
.downscale_both(A_planar, B_planar, H, W, steps)
}
\description{
Compute downscaled images, return both originals and downscaled with dims
}
\keyword{internal}
