% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.exact_cost_and_solve}
\alias{.exact_cost_and_solve}
\title{Exact pixel-level: returns \strong{1-based} assignment (A->B)}
\usage{
.exact_cost_and_solve(
  A_planar,
  B_planar,
  H,
  W,
  alpha = 1,
  beta = 0,
  method = "jv",
  maximize = FALSE
)
}
\description{
Exact pixel-level: returns \strong{1-based} assignment (A->B)
}
\keyword{internal}
