% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.expand_patch_assignment}
\alias{.expand_patch_assignment}
\title{Expand patch assignment to pixel assignment (simple spatial truncation)}
\usage{
.expand_patch_assignment(patch_assign, patches_a, patches_b, N)
}
\description{
Expand patch assignment to pixel assignment (simple spatial truncation)
}
\keyword{internal}
