% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.from_planar_rgb}
\alias{.from_planar_rgb}
\title{Convert planar RGB vector (RRR...GGG...BBB...) back to H x W x 3 array}
\usage{
.from_planar_rgb(planar, H, W)

.from_planar_rgb(planar, H, W)
}
\description{
Convert planar RGB vector (RRR...GGG...BBB...) back to H x W x 3 array

Convert planar format back to H x W x 3 array
}
\keyword{internal}
