% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.lap_assign}
\alias{.lap_assign}
\title{Solve LAP and return a consistent \strong{0-based} column index per row}
\usage{
.lap_assign(C, method = "jv", maximize = FALSE)
}
\description{
Solve LAP and return a consistent \strong{0-based} column index per row
}
\keyword{internal}
