% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.patch_cost_matrix}
\alias{.patch_cost_matrix}
\title{Compute cost matrix between two sets of patches (color+spatial)}
\usage{
.patch_cost_matrix(
  patches_a,
  patches_b,
  alpha = 1,
  beta = 0.1,
  H = NULL,
  W = NULL
)
}
\description{
Compute cost matrix between two sets of patches (color+spatial)
}
\keyword{internal}
