% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.to_array_rgb}
\alias{.to_array_rgb}
\title{Convert magick image to H x W x 3 integer array in 0–255}
\usage{
.to_array_rgb(img)
}
\description{
Convert magick image to H x W x 3 integer array in 0–255
}
\keyword{internal}
