% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.to_planar_rgb}
\alias{.to_planar_rgb}
\title{Convert H x W x 3 array to planar format (RRR...GGG...BBB...)}
\usage{
.to_planar_rgb(rgb_arr)
}
\description{
Convert H x W x 3 array to planar format (RRR...GGG...BBB...)
}
\keyword{internal}
