% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_utils.R
\name{.upscale_assignment}
\alias{.upscale_assignment}
\title{Upscale a downscaled assignment back to original resolution}
\usage{
.upscale_assignment(assign_s, H, W, Hs, Ws)
}
\description{
Upscale a downscaled assignment back to original resolution
}
\keyword{internal}
