% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_utils.R
\name{extract_matching_vars}
\alias{extract_matching_vars}
\title{Extract matching variables from data frame}
\usage{
extract_matching_vars(df, vars)
}
\description{
Extract matching variables from data frame
}
\keyword{internal}
