% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_blocks.R
\name{filter_blocks}
\alias{filter_blocks}
\title{Filter blocks based on size and balance criteria}
\usage{
filter_blocks(
  left,
  right,
  min_left,
  min_right,
  drop_imbalanced,
  imbalance_threshold
)
}
\description{
Filter blocks based on size and balance criteria
}
\keyword{internal}
