% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_core.R
\name{greedy_couples_blocked}
\alias{greedy_couples_blocked}
\title{Greedy matching with blocking}
\usage{
greedy_couples_blocked(
  left,
  right,
  left_ids,
  right_ids,
  block_col,
  vars,
  distance,
  weights,
  scale,
  max_distance,
  calipers,
  strategy
)
}
\description{
Greedy matching with blocking
}
\keyword{internal}
