% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_core.R
\name{greedy_couples_single}
\alias{greedy_couples_single}
\title{Greedy matching without blocking}
\usage{
greedy_couples_single(
  left,
  right,
  left_ids,
  right_ids,
  vars,
  distance,
  weights,
  scale,
  max_distance,
  calipers,
  strategy
)
}
\description{
Greedy matching without blocking
}
\keyword{internal}
