% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_constraints.R
\name{mark_forbidden_pairs}
\alias{mark_forbidden_pairs}
\title{Mark forbidden pairs}
\usage{
mark_forbidden_pairs(cost_matrix, forbidden_indices)
}
\description{
Generic function to mark specific pairs as forbidden.
}
\keyword{internal}
