% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_core.R
\name{match_couples_blocked}
\alias{match_couples_blocked}
\title{Match with blocking (multiple problems)}
\usage{
match_couples_blocked(
  left,
  right,
  left_ids,
  right_ids,
  block_col,
  vars,
  distance,
  weights,
  scale,
  max_distance,
  calipers,
  method
)
}
\description{
Match with blocking (multiple problems)
}
\keyword{internal}
