% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_core.R
\name{match_couples_single}
\alias{match_couples_single}
\title{Match without blocking (single problem)}
\usage{
match_couples_single(
  left,
  right,
  left_ids,
  right_ids,
  vars,
  distance,
  weights,
  scale,
  max_distance,
  calipers,
  method
)
}
\description{
Match without blocking (single problem)
}
\keyword{internal}
