% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_preprocessing.R
\name{print.variable_health}
\alias{print.variable_health}
\title{Print method for variable health}
\usage{
\method{print}{variable_health}(x, ...)
}
\arguments{
\item{x}{A variable_health object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for variable health
}
