% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_utils.R
\name{validate_matching_inputs}
\alias{validate_matching_inputs}
\title{Validate matching inputs}
\usage{
validate_matching_inputs(left, right, vars = NULL)
}
\description{
Validate matching inputs
}
\keyword{internal}
